Message_write-file
This project allows the usre to send any kind of message in there own personal computer at any file destination. First selete the file destination or the file name. The next step is to enter the message you wanted to print in the file. This message can be in any format !